+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-finance
 
-## (IN PROGRESS)
+## 5.0.0 (IN PROGRESS)
 
 * A user cannot type anything in the list of filter options for the "Status" field in the fund form. Refs UIF-458.
 * Filter in "Transfer from" and "Transfer to" fields does not work when creating a new fund. Refs UIF-457.
@@ -8,6 +8,7 @@
 * Block test rollover once rollover has been done for given ledger and FYs. Refs UIF-465.
 * Update Node.js to v18 in GitHub Actions. Refs UIF-471.
 * Need new permission(s) to view Finance settings in UI. Refs UIF-442.
+* *BREAKING* bump `react-intl` to `v6.4.4`. Refs UIF-472.
 
 ## [4.0.4](https://github.com/folio-org/ui-finance/tree/v4.0.4) (2023-03-31)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v4.0.3...v4.0.4)